+++ pkg/controller/elasticsearch.go
@@ -6,24 +6,28 @@ import (
 
 	"github.com/appscode/go/encoding/json/types"
 	"github.com/appscode/go/log"
-	"github.com/kubedb/apimachinery/apis"
-	api "github.com/kubedb/apimachinery/apis/kubedb/v1alpha1"
-	"github.com/kubedb/apimachinery/client/clientset/versioned/typed/kubedb/v1alpha1/util"
-	"github.com/kubedb/apimachinery/pkg/eventer"
-	validator "github.com/kubedb/elasticsearch/pkg/admission"
 	"github.com/pkg/errors"
+	appsv1 "k8s.io/api/apps/v1"
 	core "k8s.io/api/core/v1"
+	policyv1beta1 "k8s.io/api/policy/v1beta1"
 	kerr "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/util/intstr"
 	clientsetscheme "k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/tools/reference"
 	kutil "kmodules.xyz/client-go"
 	core_util "kmodules.xyz/client-go/core/v1"
 	dynamic_util "kmodules.xyz/client-go/dynamic"
 	meta_util "kmodules.xyz/client-go/meta"
+	policy_util "kmodules.xyz/client-go/policy/v1beta1"
 	storage "kmodules.xyz/objectstore-api/osm"
+	"kubedb.dev/apimachinery/apis"
+	api "kubedb.dev/apimachinery/apis/kubedb/v1alpha1"
+	"kubedb.dev/apimachinery/client/clientset/versioned/typed/kubedb/v1alpha1/util"
+	"kubedb.dev/apimachinery/pkg/eventer"
+	validator "kubedb.dev/elasticsearch/pkg/admission"
 )
 
 func (c *Controller) create(elasticsearch *api.Elasticsearch) error {
@@ -90,29 +94,42 @@ func (c *Controller) create(elasticsearch *api.Elasticsearch) error {
 		)
 	}
 
+	// ensure appbinding before ensuring Restic scheduler and restore
+	_, err = c.ensureAppBinding(elasticsearch)
+	if err != nil {
+		log.Errorln(err)
+		return err
+	}
+
 	if _, err := meta_util.GetString(elasticsearch.Annotations, api.AnnotationInitialized); err == kutil.ErrNotFound &&
 		elasticsearch.Spec.Init != nil &&
-		elasticsearch.Spec.Init.SnapshotSource != nil {
-
-		snapshotSource := elasticsearch.Spec.Init.SnapshotSource
+		(elasticsearch.Spec.Init.SnapshotSource != nil || elasticsearch.Spec.Init.StashRestoreSession != nil) {
 
 		if elasticsearch.Status.Phase == api.DatabasePhaseInitializing {
 			return nil
 		}
 
-		jobName := fmt.Sprintf("%s-%s", api.DatabaseNamePrefix, snapshotSource.Name)
-		if _, err := c.Client.BatchV1().Jobs(snapshotSource.Namespace).Get(jobName, metav1.GetOptions{}); err != nil {
-			if !kerr.IsNotFound(err) {
-				return err
+		// add phase that database is being initialized
+		mg, err := util.UpdateElasticsearchStatus(c.ExtClient.KubedbV1alpha1(), elasticsearch, func(in *api.ElasticsearchStatus) *api.ElasticsearchStatus {
+			in.Phase = api.DatabasePhaseInitializing
+			return in
+		}, apis.EnableStatusSubresource)
+		if err != nil {
+			return err
+		}
+		elasticsearch.Status = mg.Status
+
+		init := elasticsearch.Spec.Init
+		if init.SnapshotSource != nil {
+			err = c.initializeFromSnapshot(elasticsearch)
+			if err != nil {
+				return fmt.Errorf("failed to complete initialization. Reason: %v", err)
 			}
-		} else {
+			return err
+		} else if init.StashRestoreSession != nil {
+			log.Debugf("Elasticsearch %v/%v is waiting for restoreSession to be succeeded", elasticsearch.Namespace, elasticsearch.Name)
 			return nil
 		}
-		err = c.initialize(elasticsearch)
-		if err != nil {
-			return fmt.Errorf(`failed to complete initialization for "%v/%v". Reason: %v`, elasticsearch.Namespace, elasticsearch.Name, err)
-		}
-		return nil
 	}
 
 	es, err := util.UpdateElasticsearchStatus(c.ExtClient.KubedbV1alpha1(), elasticsearch, func(in *api.ElasticsearchStatus) *api.ElasticsearchStatus {
@@ -162,11 +179,6 @@ func (c *Controller) create(elasticsearch *api.Elasticsearch) error {
 		return nil
 	}
 
-	_, err = c.ensureAppBinding(elasticsearch)
-	if err != nil {
-		log.Errorln(err)
-		return err
-	}
 	return nil
 }
 
@@ -239,17 +251,17 @@ func (c *Controller) ensureBackupScheduler(elasticsearch *api.Elasticsearch) err
 	return nil
 }
 
-func (c *Controller) initialize(elasticsearch *api.Elasticsearch) error {
-	es, err := util.UpdateElasticsearchStatus(c.ExtClient.KubedbV1alpha1(), elasticsearch, func(in *api.ElasticsearchStatus) *api.ElasticsearchStatus {
-		in.Phase = api.DatabasePhaseInitializing
-		return in
-	}, apis.EnableStatusSubresource)
-	if err != nil {
-		return err
-	}
-	elasticsearch.Status = es.Status
-
+func (c *Controller) initializeFromSnapshot(elasticsearch *api.Elasticsearch) error {
 	snapshotSource := elasticsearch.Spec.Init.SnapshotSource
+	jobName := fmt.Sprintf("%s-%s", api.DatabaseNamePrefix, snapshotSource.Name)
+	if _, err := c.Client.BatchV1().Jobs(snapshotSource.Namespace).Get(jobName, metav1.GetOptions{}); err != nil {
+		if !kerr.IsNotFound(err) {
+			return err
+		}
+	} else {
+		return nil
+	}
+
 	// Event for notification that kubernetes objects are creating
 	c.recorder.Eventf(
 		elasticsearch,
@@ -444,8 +456,35 @@ func (c *Controller) UpsertDatabaseAnnotation(meta metav1.ObjectMeta, annotation
 	}
 
 	_, _, err = util.PatchElasticsearch(c.ExtClient.KubedbV1alpha1(), elasticsearch, func(in *api.Elasticsearch) *api.Elasticsearch {
-		in.Annotations = core_util.UpsertMap(elasticsearch.Annotations, annotation)
+		in.Annotations = core_util.UpsertMap(in.Annotations, annotation)
 		return in
 	})
+	return err
+}
+
+func (c *Controller) createPodDisruptionBudget(sts *appsv1.StatefulSet, maxUnavailable *intstr.IntOrString) error {
+	ref, err := reference.GetReference(clientsetscheme.Scheme, sts)
+	if err != nil {
+		return err
+	}
+
+	m := metav1.ObjectMeta{
+		Name:      sts.Name,
+		Namespace: sts.Namespace,
+	}
+	_, _, err = policy_util.CreateOrPatchPodDisruptionBudget(c.Client, m,
+		func(in *policyv1beta1.PodDisruptionBudget) *policyv1beta1.PodDisruptionBudget {
+			in.Labels = sts.Labels
+			core_util.EnsureOwnerReference(&in.ObjectMeta, ref)
+
+			in.Spec.Selector = &metav1.LabelSelector{
+				MatchLabels: sts.Spec.Template.Labels,
+			}
+
+			in.Spec.MaxUnavailable = maxUnavailable
+
+			in.Spec.MinAvailable = nil
+			return in
+		})
 	return err
 }