+++ test/e2e/framework/elasticsearch.go
@@ -2,24 +2,32 @@ package framework
 
 import (
 	"fmt"
+	"strconv"
 	"time"
 
 	"github.com/appscode/go/crypto/rand"
 	jtypes "github.com/appscode/go/encoding/json/types"
 	"github.com/appscode/go/types"
-	api "github.com/kubedb/apimachinery/apis/kubedb/v1alpha1"
-	util "github.com/kubedb/apimachinery/client/clientset/versioned/typed/kubedb/v1alpha1/util"
-	"github.com/kubedb/elasticsearch/pkg/util/es"
 	. "github.com/onsi/gomega"
 	core "k8s.io/api/core/v1"
+	policy "k8s.io/api/policy/v1beta1"
 	kerr "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	meta_util "kmodules.xyz/client-go/meta"
+	api "kubedb.dev/apimachinery/apis/kubedb/v1alpha1"
+	util "kubedb.dev/apimachinery/client/clientset/versioned/typed/kubedb/v1alpha1/util"
+	"kubedb.dev/elasticsearch/pkg/util/es"
 )
 
 var (
 	JobPvcStorageSize = "2Gi"
 	DBPvcStorageSize  = "1Gi"
+)
+
+const (
+	kindEviction = "Eviction"
 )
 
 func (i *Invocation) CombinedElasticsearch() *api.Elasticsearch {
@@ -102,31 +110,31 @@ func (i *Invocation) DedicatedElasticsearch() *api.Elasticsearch {
 }
 
 func (f *Framework) CreateElasticsearch(obj *api.Elasticsearch) error {
-	_, err := f.extClient.KubedbV1alpha1().Elasticsearches(obj.Namespace).Create(obj)
+	_, err := f.dbClient.KubedbV1alpha1().Elasticsearches(obj.Namespace).Create(obj)
 	return err
 }
 
 func (f *Framework) GetElasticsearch(meta metav1.ObjectMeta) (*api.Elasticsearch, error) {
-	return f.extClient.KubedbV1alpha1().Elasticsearches(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
+	return f.dbClient.KubedbV1alpha1().Elasticsearches(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
 }
 
 func (f *Framework) TryPatchElasticsearch(meta metav1.ObjectMeta, transform func(*api.Elasticsearch) *api.Elasticsearch) (*api.Elasticsearch, error) {
-	elasticsearch, err := f.extClient.KubedbV1alpha1().Elasticsearches(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
+	elasticsearch, err := f.dbClient.KubedbV1alpha1().Elasticsearches(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
 	if err != nil {
 		return nil, err
 	}
-	elasticsearch, _, err = util.PatchElasticsearch(f.extClient.KubedbV1alpha1(), elasticsearch, transform)
+	elasticsearch, _, err = util.PatchElasticsearch(f.dbClient.KubedbV1alpha1(), elasticsearch, transform)
 	return elasticsearch, err
 }
 
 func (f *Framework) DeleteElasticsearch(meta metav1.ObjectMeta) error {
-	return f.extClient.KubedbV1alpha1().Elasticsearches(meta.Namespace).Delete(meta.Name, nil)
+	return f.dbClient.KubedbV1alpha1().Elasticsearches(meta.Namespace).Delete(meta.Name, nil)
 }
 
 func (f *Framework) EventuallyElasticsearch(meta metav1.ObjectMeta) GomegaAsyncAssertion {
 	return Eventually(
 		func() bool {
-			_, err := f.extClient.KubedbV1alpha1().Elasticsearches(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
+			_, err := f.dbClient.KubedbV1alpha1().Elasticsearches(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
 			if err != nil {
 				if kerr.IsNotFound(err) {
 					return false
@@ -141,10 +149,22 @@ func (f *Framework) EventuallyElasticsearch(meta metav1.ObjectMeta) GomegaAsyncA
 	)
 }
 
+func (f *Framework) EventuallyElasticsearchPhase(meta metav1.ObjectMeta) GomegaAsyncAssertion {
+	return Eventually(
+		func() api.DatabasePhase {
+			db, err := f.dbClient.KubedbV1alpha1().Elasticsearches(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
+			Expect(err).NotTo(HaveOccurred())
+			return db.Status.Phase
+		},
+		time.Minute*5,
+		time.Second*5,
+	)
+}
+
 func (f *Framework) EventuallyElasticsearchRunning(meta metav1.ObjectMeta) GomegaAsyncAssertion {
 	return Eventually(
 		func() bool {
-			elasticsearch, err := f.extClient.KubedbV1alpha1().Elasticsearches(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
+			elasticsearch, err := f.dbClient.KubedbV1alpha1().Elasticsearches(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
 			Expect(err).NotTo(HaveOccurred())
 			return elasticsearch.Status.Phase == api.DatabasePhaseRunning
 		},
@@ -184,12 +204,12 @@ func (f *Framework) EventuallyElasticsearchIndicesCount(client es.ESClient) Gome
 }
 
 func (f *Framework) CleanElasticsearch() {
-	elasticsearchList, err := f.extClient.KubedbV1alpha1().Elasticsearches(f.namespace).List(metav1.ListOptions{})
+	elasticsearchList, err := f.dbClient.KubedbV1alpha1().Elasticsearches(f.namespace).List(metav1.ListOptions{})
 	if err != nil {
 		return
 	}
 	for _, e := range elasticsearchList.Items {
-		if _, _, err := util.PatchElasticsearch(f.extClient.KubedbV1alpha1(), &e, func(in *api.Elasticsearch) *api.Elasticsearch {
+		if _, _, err := util.PatchElasticsearch(f.dbClient.KubedbV1alpha1(), &e, func(in *api.Elasticsearch) *api.Elasticsearch {
 			in.ObjectMeta.Finalizers = nil
 			in.Spec.TerminationPolicy = api.TerminationPolicyWipeOut
 			return in
@@ -197,7 +217,67 @@ func (f *Framework) CleanElasticsearch() {
 			fmt.Printf("error Patching Elasticsearch. error: %v", err)
 		}
 	}
-	if err := f.extClient.KubedbV1alpha1().Elasticsearches(f.namespace).DeleteCollection(deleteInForeground(), metav1.ListOptions{}); err != nil {
+	if err := f.dbClient.KubedbV1alpha1().Elasticsearches(f.namespace).DeleteCollection(deleteInForeground(), metav1.ListOptions{}); err != nil {
 		fmt.Printf("error in deletion of Elasticsearch. Error: %v", err)
 	}
+}
+
+func (f *Framework) EvictPodsFromStatefulSet(meta metav1.ObjectMeta) error {
+	var err error
+	labelSelector := labels.Set{
+		meta_util.ManagedByLabelKey: api.GenericKey,
+		api.LabelDatabaseKind:       api.ResourceKindElasticsearch,
+		api.LabelDatabaseName:       meta.GetName(),
+	}
+	// get sts in the namespace
+	stsList, err := f.kubeClient.AppsV1().StatefulSets(meta.Namespace).List(metav1.ListOptions{LabelSelector: labelSelector.String()})
+	if err != nil {
+		return err
+	}
+	for _, sts := range stsList.Items {
+		// if PDB is not found, send error
+		var pdb *policy.PodDisruptionBudget
+		pdb, err = f.kubeClient.PolicyV1beta1().PodDisruptionBudgets(sts.Namespace).Get(sts.Name, metav1.GetOptions{})
+		if err != nil {
+			return err
+		}
+		eviction := &policy.Eviction{
+			TypeMeta: metav1.TypeMeta{
+				APIVersion: policy.SchemeGroupVersion.String(),
+				Kind:       kindEviction,
+			},
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      sts.Name,
+				Namespace: sts.Namespace,
+			},
+			DeleteOptions: &metav1.DeleteOptions{},
+		}
+
+		if pdb.Spec.MaxUnavailable == nil {
+			return fmt.Errorf("found pdb %s spec.maxUnavailable nil", pdb.Name)
+		}
+
+		// try to evict as many pod as allowed in pdb. No err should occur
+		maxUnavailable := pdb.Spec.MaxUnavailable.IntValue()
+		for i := 0; i < maxUnavailable; i++ {
+			eviction.Name = sts.Name + "-" + strconv.Itoa(i)
+
+			err := f.kubeClient.PolicyV1beta1().Evictions(eviction.Namespace).Evict(eviction)
+			if err != nil {
+				return err
+			}
+		}
+
+		// try to evict one extra pod. TooManyRequests err should occur
+		eviction.Name = sts.Name + "-" + strconv.Itoa(maxUnavailable)
+		err = f.kubeClient.PolicyV1beta1().Evictions(eviction.Namespace).Evict(eviction)
+		if kerr.IsTooManyRequests(err) {
+			err = nil
+		} else if err != nil {
+			return err
+		} else {
+			return fmt.Errorf("expected pod %s/%s to be not evicted due to pdb %s", sts.Namespace, eviction.Name, pdb.Name)
+		}
+	}
+	return err
 }